Dovahscript: miscellaneous porting work
Ported over the Lua `benchmark` class. There are improvements we can make once we've fully migrated to the new script engine, when script compatibility between the two (ironically for benchmarking's sake) is no longer a concern.

Fixed up `push_native_object` so that we'll be able to reliably handle resource objects.

Work continues on porting the `dovah` Lua library.