Dovahscript: `dovahscript::send_script_task` and friends
Added helper functions to the `dovahscript` namespace: `send_script_task` and `send_script_ui_task`. The purpose of these functions is to make it so that we don't need absolutely every API file to include `coordinator.h`; hopefully now we can avoid having to recompile basically the entire script engine every time we make any changes to the coordinator subsystem.