Dovahscript: fixed lifetime handling for unreferenced model observers
Added a single constexpr bool that we can use to force-enable all of our various constexpr logging bools in Dovahscript.

Fixed lifetime handling for model observers that become unreferenced; mistakes in the lifetime check queue and the hierarchy crawler have been corrected, and as a result model observers are now able to keep their models' associated widgets alive (which keeps the models alive, which keeps the observers alive).