Dovahscript: wrapper `__close` fixes; ported QLabel; Lua helper fix
Fixed the helper functions for dumping Lua variables to a string: they were calling `luaL_tolstring` with a negative index in some places, but that function doesn't run `lua_absindex` and so mishandles those indices. This mainly affected the printing of tables and userdata, where the memory address printed would be for the table's own `__name` string value and not for the table itself.

Ported over the `text` UI widget.

Fixed a bug in how we tear down wrappers during `__close`. We need special handling in `core::subsystems::userdata::on_wrapper_destroyed`, as described in the comments there.