Dovahscript: implemented "eval scripts"
The event registration helpers for the event subsystem no longer include `coordinator.h`.

Added the abilty to run an "eval script" while a script session is active. Essentially, frontends can offer a "console" textbox and while a script is active (i.e. it has windows open), users can enter code and queue it to run as soon as possible; this is a useful tool for debugging a wayward script.

When scripts run, the script code textbox is now readonly instead of disabled.