Dovahscript: ported line, spinbox
Ported the `line` and `spinbox` APIs. The former required the creation of a new QFrame subclass, DovahscriptLineWidget, since we no longer have the ability to override widgets' metatables. (Really, we never could; the previous script engine used widget properties to store metatable overrides, but those can't be accessed thread-safely; they may be reentrant, but Qt can create and modify properties of its own for state-keeping, so in practice it's 100% unsafe.)