Dovahscript: ported radio_button, radio_group; started DKDoubleSlider
Ported over the `radio_button` and `radio_group` script APIs.

Attempted to start work on DKDoubleSlider, with the intention of making a QSlider variant whose values are doubles instead of ints. However, it's looking like that'll be a LOT more work than I expected, so I've excluded it from the VS project and I'll shelve it for now.