Dovahscript: hierarchy bridge lifetime fixes; radio button API fixes
Fixed an event name typo in the `radio_group` event registrar.

Fixed a mistake in `hierarchy_crawler` that caused us to mishandle hierarchy bridges, such that an unreferenced object in a bridged-to hierarchy could not only be deleted, but also cause the deletion of the bridge even if the bridge were referenced.

The `lifetime_check_queue` now recognizes QButtonGroups as hierarchy objects.

Added documentation comments to `radio_button`'s `id` field.