Dovahscript: ported canvas-related APIs; lifetime fixes
Ported over the Lua API classes for CanvasWidget.

The lifetime subsystem now explicitly keeps a list of `DovahscriptCanvasWidgetLayerData` rather than just generic `CanvasWidgetLayerData`. The `DovahscriptCanvasWidgetLayerDataResource` class now subclasses the former instead of the latter; though we've designed our APIs so that access to it is transparent (i.e. you can't refer directly to a DCWLDR, but instead only to its resource), I want to ensure that any potential changes to the API design in the future don't break anything.

Fixed the lifetime queue so that when checking canvas widget layer data, it checks whether the data is Lua-referenced. This is needed for lifetime checks resulting from a (possibly still Lua-referenced) layer data being unhooked from all layers.

The lifetime system can now be made aware of newly-created non-hierarchy objects, such as layer data.