Dovahscript: API fixes; vital task system fixes
Fixed various compiler errors for the form APIs, as well as a mistake which prevented the benchmarking APIs from working. The program now builds.

Discovered and attempted to remedy a nasty oversight in the cross-thread task system that could cause access violations or double-frees when aborting a script early (i.e. while a blocking task was in the queue, waiting to be processed). I've documented the problem and attempted to fix it; the fix is not yet tested.

When making modifications to forms, we have to signal the main thread before and after the changes are made, and ensure that the main thread fully responds to those signals; else we might break parts of the frontend. Accordingly, this relies on cross-thread tasks under the hood. The above change to the task system creates the risk that the "before" task could be sent while the "after" task is not; accordingly, changes have been made to these tasks, with integration with the `coordinator`, to ensure that forms are never left in a "limbo" state where the frontend is perpetually waiting for modifications to finish.

Removed the `fire_and_forget` option from all cross-thread task objects; there is no sane situation in which a non-blocking task would be non-fire-and-forget. Fire-and-forget blocking tasks would be nice but are not essential.