Dovahscript: major resource optimization
Fixed the `coordinator` destructor to prevent a potential crash-on-exit -- the result of an assertion failure stemming from us running a double-teardown. The `coordinator` destructor doesn't need to trigger teardown, in part because the coordinator should only be destroyed when the entire program exits (no need to manually free everything; the OS will wipe the slate clean for us anyway).

DovahscriptResources now use QImage for the client-thread data, significantly alleviating overhead associated with creating and destroying QPixmaps; additionally, when resynchronizing an image whose size and format have not changed, we use QPainter (with composition mode "source," which acts as a "replace" mode) instead of assign-and-detach, to avoid deletions and allocations that the latter would trigger.

The resource optimization has had a spectacular effect on the performance of our heightmap script. In Release, small and medium workloads (Blackreach, Sovngarde) now perform only marginally slower than the old engine, while massive workloads (Tamriel) are actually twice as fast.

At this point, I think I can say we've achieved parity-or-better with the old engine, and we've accomplished all the improvements we needed for the engine core.