Started work on script packages
It is now possible to create "script packages" consisting of an entire folder full of Lua files meant to be run as a unit; it's also now possible to actually select and run script packages in the editor.

For testing purposes, I took the heightmap script and created a multi-file script package out of it.

We also have a `permission_set` struct, in preparation for implementing the permissions subsystem.