Relicense as GPLv3, with CLA
I'm relicensing DovahKIt from CC-BY-NC-SA to GPLv3.

Added a guide to contributing to the project, with advice on how to get things ready in Visual Studio.

Added a Contributor License Agreement in a linked GitHub Gist. The CLA just takes a clause already in GitHub's Terms of Service (when you contribute to a repo, your contribution goes under the same license as the repo) and makes it explicit, so I don't have to rely on GitHub never changing that clause out from under me.