ColorPickerButton fix; started work on Qt-based INI code
ColorPickerButton should now work properly when its signals are blocked.

Started work on new INI-handling code, with the same design goals as the original but built on Qt. This code should be more appropriate for implementing settings within DovahKit, but I think it may also be viable to use it for loading Skyrim.ini and friends; we'll see.