Work on INI helpers (loading, saving, etc.)
Files now keep track of their containing Settings without relying on the QObject hierarchy.

INI settings now store a qMetaType ID, and use this when their serialization type is `String`. This is used for conversions, to allow struct values to be seamlessly written provided those structs are convertible to and from QString (and Qt has been made aware of this).

Set a Setting's current value only emits a valueChanged signal if the value being written actually differs from what was there previously.

Wrote code to load and save INI files, loosely porting some of the logic of the old system. Currently untested.

Added `File::categoryNameCanonicalCase` which normalizes the case on a category name.