Work on INI helpers
The helpers for setting-edit controls now block signals when updating a control in response to the pending value being discarded (which can happen if e.g. the file is reloaded from disk).

Added templated `Setting::setCurrentValue` and `Setting::setPendingValue` overloads to avoid the need to constantly wrap struct arguments in `QVariant::fromValue<T>(...)`.

Added a templated `Setting::modifyPendingValue` member function in order to make it easier to write UI code for editing the values of struct-settings, where individual controls may edit different fields on the struct.

Added `Setting::hasPendingValue` and `Setting::pendingOrCurrentValue` member functions. The former is not strictly necessary but should serve as a guide, of sorts, on how values "work."

Fixed `File::save` so that if you choose to preserve formatting but there's no already-existing file, we properly create a new file from scratch instead of just silently failing.

Added a test struct to our INI helper test code, along with a setting that uses that struct type as its value.