INI helper work; script textarea API; cleanup and maintenance
Moved `QLuaCheckbox` to `DovahscriptCheckbox` and updated the `checkbox` script API accordingly.

Created `DovahscriptTextarea` and a `textarea` script API for multi-line textboxes.

Unfinished Lua classes and backend loaded-form classes now static-assert on a constexpr bool instead of using #pragma message, which was unreliable. We can change the constexpr bool to false when we think we're ready to ship in order to double-check our work.

Renamed events on the `textbox` API: OnKeyPressed is now OnChanged, and OnChanged is now AfterChanged.

INI helpers:

Wrote usage examples and improved documentation comments.

`Setting::setPendingValue` now fires the pending-value-discarded signal if the passed-in QVariant is invalid.

The "binding" helpers now take advantage of `Setting::pendingOrCurrentValue()`.