Dovahscript: work on file access APIs
Added a `binary_view` class meant to be similar to JavaScript's `DataView`. This will be used for any resources that are loaded as binary, and should now replace the `unknown` resource class.

Started work on a `load_and_push_resource` API helper, which will be used to load files from the game assets or from a script package's folder.

Separately, backend fixes for us removing the `endianness.h` helper header in favor of a wholly C++20 header.