Dovahscript: color picker button; events refactor; file loading APIs
Created `cobb::class_list` as a way of metaprogramming a loop that runs on a set of classes.

Refactored parts of the `events` subsystem in Dovahscript to make use of `class_list`, making the code for handling events significantly cleaner. As part of this refactor, we now loop over an event target's QMetaObjects in ascending order; this allows us to avoid superclass/subclass confusion when unique events exist for both a superclass and a subclass.

Added a `color_button` widget to Lua.

Added a `parent` getter to Lua widgets.

Added a (not yet tested) `dovah.package.load_file` API.

The `ColorPickerButton::colorChanged` signal now includes the new QColor as an argument.

Started work on a new script package that is intended to let users map out the locations of all references of given base forms in a given worldspace.