Dovahscript: binary view and package file API tests and fixes
Fixed the API for loading resources from a script package; internally, it now properly indicates that it wants a package asset and not a game asset.

Tested the API for loading resources, and made fixes to the functions that do quick correctness checks for JPEGs and PNGs.

Made several fixes to the `binary_view` APIs.