Work on DKCollapsiblePane: made a Qt Designer plug-in
I've created a Qt Designer plug-in (for the first time!) to enable the use of more complex widgets inside of Qt Designer, particularly in cases where the "promoted widget" feature would not allow for correct editing behavior. DKCollapsiblePane is one such case.

Fixed the visual styling on DKCollapsiblePane's header, and fixed the widget's sizing when its body is collapsed.

As a necessary change to make it usable in Qt Designer, DKCollapsiblePane now allows you to replace its viewport a la QScrollArea.