Dovahscript: ability to browse for a script package; event CTD fix
Fixed an oversight in the events subsystem which led to a crash when a script finished executing. Optimized the events subsystem's teardown code.

The script package UI now features a button allowing users to browse for a script package in any folder. If the selected folder doesn't have a valid manifest.xml file, the error is reported to the user. This should help with cases where a manifest file in the normal folder is invalid (there's no UI flow to report that).

Because we can now browse for script packages in any folder, I've moved the debug packages to a subfolder so that they don't automatically show up in the in-program listings.