Dovahscript: work on REFR; minor tweaks to script manifest parsing
Started work on Lua wrappers for REFR forms and their members. I've not tested any of them yet, but the base form and form flags should be accessible, the position should be readable, and the rotation should be accessible.

Modified the script package manifest parser to accept HTML entities.

Modified `cobb::qt::minimize_indent` to also strip out empty newlines at the start and end of a string. This will help for multi-line text strings in tags that have been line-broken and indented.