Dovahscript: improvements to warning and error reporting
The `widget.layout_margins` setter now warns if called on a widget with no current layout. Unfortunately, that means it is now a blocking API, but scripts should not be running it especially often, so that's something I can accept.

When the coordinator subsystem passes script filenames into Lua, it now prefixes them with '=' so that they're treated as literal chunknames. This is the difference between `[string "userscript"]` and `[userscript]`.

When Dovahscript reports warnings or errors, it now uses a custom traceback function that is explicitly designed to accommodate our class system: warnings and errors emitted from inside of a class getter or setter should now report the name of the class and its getter or setter, instead of the stack frame merely displaying as `[C]: ?`. Unfortunately, however, we are currently unable to account for the custom message text supplied by Lua's "argcheck" and "argerror" functions: these lookup the function name on their own, and so while the stack trace properly shows the class member name, the error text is still `bad argument #2 to ? (custom text here)`.

To facilitate the above change, I have altered how our class system's `__index` and `__newindex` functions call the underlying getter or setter. Previously I set it up so that the only things left on the Lua stack would be the getter/setter function and its arguments (including the self-object), such that after the call, the stack would be empty. However, our improved error reporting relies on being able to reach into the __(new)index stack frame and query its local variables (which, in practice, is just its Lua stack), and it seems that `lua_call` removes/transfers arguments out of the stack before the function is actually called. As such, __(new)index now sets up the stack so that the self-object is redundantly added before the call is set up and made; this ensures that there is a single local variable accessible to our debug code, holding the self-object.