Dovahscript: canvas export fixes
When the user saves a canvas as a ZIP archive of its layers, the layer filenames are now consistent with the layer numbers listed in the `info.txt` file.

The `info.txt` file now properly prints information for layers. Previously it didn't, because I'd made a typo and MSVC somehow failed to emit a compiler error for passing `A*` to a function that expected `A&`.