Dovahscript benchmark; constexpr form types; DKFormListPane; LTEX UI
Rearranged Dovahscript's `benchmark` class, for consistency with other classes, and modified existing scripts to match.

Moved `dovah::form_type` and `dovah::form_type_info` into their own file, and made them constexpr. That file is included from core.h. A separate file is used because constexpr stuff has to be included in the header in full, which is... as annoying as it sounds.

Modified the `open_window_for_form` function to use an array mapping form types to dialog constructors; this will make the list of form type windows easier to maintain.

Created a new plug-in-based widget: DKFormListPane. This is intended as a replacement to FormListListview and friends, and is cleaner and more powerful.

Created the UI for editing LTEX (LandTexture) forms.