TextureSet and BSA file picker work
Built a TextureSet editing dialog in Qt Designer, though it currently has no C++ code backing it.

Started work on a collection of widgets and classes intended to allow users to select file paths from the loaded BSAs; that is, I want to offer an "Open File" dialog for archived files, along with the option to fall back to a real "Open File" dialog for picking loose files. When this is all done, we can use it in the TextureSet dialog, and we may be able to go back to a few older dialogs (e.g. worldspaces) and use it there as well.