DKGameFilePicker is now functional and in use
The DKGameFilePicker is now fully-functional and is used in the Cell dialog for CELL/XWEM. It has also been added to the Qt Designer plug-in.

DKBSABrowseDialog now sets the proper tab order on its contents, and fixes were made to `offerLooseFile` (to work around QDir being an inconsistently-designed API, and to work around QFileDialog limits) and to various other functions that handle paths.

The editor core now creates a DKBSACollectionModelBackend and populates it as appropriate.