BSA widgets hardened against file path issues
DKBSABrowseDialog can now validate file paths and display warnings when the user picks a path that could cause issues in the game engine.

DKGameFilePicker can now be given ValidationOptions, to be used by DKBSABrowseDialog to check for specific problem cases.

Issues in Bethesda's path handling are now documented. Information is limited by the fact that I don't know when or for which assets each documented path function is used.