Work on the TextureSet UI; TXST/DODT backend changes
Added `cobb::unreachable` so that I don't have `__assume`s just hangin' out everywhere, and replaced the latter with the former.

TextureSet loaded-forms now store a pointer to a heap-allocated decal data structure. This is how the game itself does things; the decal data struct is heap-allocated rather than existing on the TXST, and decal data is only created if a DODT subrecord is loaded; the "Decal data" checkbox in the Creation Kit just determines if the struct exists / the subrecord is present.

Checked in the CK and set minimum/maximum values for various fields in the TextureSet UI, along with figuring out which texture paths display when.

Started writing code for the TextureSet editing window. It's blocked by the fact that I forgot to promote a QPushButton in the UI to a ColorPickerButton... but I think it'd be cleaner to just have a DKColorPickerButton widget in the Qt Designer plug-in...