DKColorPickerButton; TextureSet UI
Copied the ColorPickerButton widget to create DKColorPickerButton, and implemented it as a Qt Designer-compatible widget. Made some enhancements to how it handles its text color, too.

DKGameFilePicker now offers `effectiveStem` and `setRawPath` functions, and a `rawPathChanged` signal. The "raw path" is the path with all `PathFormat` options applied, while the plain "path" is the path that the widget maintains in its own internal format. The "effective stem" is the path stem, if one has been set, or a path stem supplied by the StandardConfiguration, if one has been set. These changes resolve a variety of issues present when using the control in a real-world setting, with the `OmitPathStem` path format.

DKCollapsiblePane is now under CC0.

The TextureSet UI now uses DKColorPickerButton.

The TextureSet UI has received significant improvements. Rows in the path table are now sized consistently, and individual cells can't be modified directly by double-clicking. The path table is now properly synchronized to the DKGameFilePicker control. The decal data section now has the checkbox properly handled when loading a form. The path table now properly generates QTableWidgetItems on load (simply setting the row and column count isn't enough; empty cells with nullptr QTableWidgetItems are valid and are what gets created).