Started work on the asset manager
Fix: TextureSet window is now known to the rest of the program, and can be opened. (This was from a previous commit, but I forgot to push it.)

Started work on the asset manager. This system will resemble Dovahscript's resource subsystem, allowing the entire program to load and reuse game assets in memory. The first test-case is offering texture previews in UI windows, but this will eventually also power the render window.