Work on assets; DKTextureAssetPane; first functional tests
DovahKitAssetHandle is no longer a simple smart pointer; rather, it's a QObject, and one that you're expected to have on hand before you request an asset. You should hook the "ready" signal on the handle before you attempt to load anything; the handle will re-fire "ready" if you request an asset that's already loaded, or otherwise forward "ready" from the asset itself.

DovahKitAssets now know their own paths (which will make multithreaded loading easier down the line), and have lifetime management implemented. Their loading state flags are now under a mutex; this is needed for DovahKitAssetHandle to be able to safely check whether/how to forward the "ready" signal.

DovahKitAssetManager has a work-in-progress worker thread struct.

Created DKTextureAssetPane, a widget that can be given an asset or asset path and display the asset if it's a texture. If the asset is still loading, an animated loading spinner is displayed; if no asset has been set, a null icon is displayed.

Created a test for DKTextureAssetPane. This verifies that the general process of retrieving assets works, at least right now while things are single-threaded for testing purposes. You can browse for an asset, pick one, and have it displayed; debug logging confirms that the previous asset (if any) is properly unloaded.

Removed unused code from DKGameFilePicker.