Asset manager fixes; multithreaded asset loading; asset unload handling
DovahKitCore's constructor makes sure that DovahKitAssetManager exists as soon as it can be safely created.

DovahKitAssetManager now uses worker threads to coordinate the loading of assets. Assets unload on the main thread. Worker threads currently initialize COM in order to make use of DirectXTex; if we figure out how to render DDS textures directly to a UI widget, we can eventually get rid of this.

DovahKitAssetReceptor now allows users to handle the case of an asset being unloaded out from under the receptor (e.g. by the asset manager when mass-abandoning game data and, by implication, BSA content).

DKTextureAssetPane now catches asset unloads and ceases to render the asset (which would be a null pointer access at that point).