Asset manager lock fixes; DKTextureAssetPane throttle, edge-case fix
Fixed locking for `DovahKitAssetManager::requestAsset`: we no longer briefly unlock when creating a new asset (if we allowed multithreaded access, that unlock would've created a brief window in which multiple assets could be created for the same key). At this point, we could plausibly make the non-form-asset overload available to multiple threads (the form-asset overload still has to be locked to the main thread unless we find a way to actually queue up the form-asset load to happen on the main thread).

DKTextureAssetPane now has a "throttle" feature which makes it so that if the asset is set several times within a brief (configurable) window, the widget only attempts to load the last provided asset.

Fixed an oversight in DKTextureAssetPane that made it so that calling `setAsset` with the asset the pane was already using would get it stuck in a loading spinner. This occurred because it would dutifully update its state (turn on the spinner and write to its receptor), but the receptor won't re-fire ready/failed signals if an assignment doesn't actually change what asset it's referring to, and the widget was counting on those signals always being fired. In practice, this produced a bug where upon opening a TextureSet editing dialog and having it default to previewing the diffuse texture, you could click the diffuse texture in the texture list (setting the preview pane to the asset it's already showing) and get the preview stuck on a loading spinner.