Various teardown fixes for assets and FormPicker
FormPicker fix: we now properly clear the picker's value onDataAbandonImminent. This fixes a crash that could occur if a FormPicker was present (and had a form selected) when data was unloaded and reloaded; it would attempt to reselect its previous (now deleted) stub.

DovahKitAssets now clear their form stub pointer when unloading, if they are form-assets. This fixes an edge-case that DovahKitAssetManager::unloadAll was tripping; refer to comments in DovahKitAsset::unload. As a consequence, the stub pointer is no longer const.

DovahKitAssetManager now has an atomic bool indicating that it's in the middle of an unloadAll call; this allows us to skip asset-unreferenced logic during teardown, which we need to do to avoid deadlocks.