Vulkan tutorial progress: triangle is rendered
We now subclass QWidget for our render window's 3D view. The subclass sets widget attributes and overrides functions as necessary to prevent Qt from painting to it normally, and uses a fine-grained timer to render Vulkan frames.

The Vulkan subsystem we've created, meanwhile, handles most of the first tutorial, up to and including rebuilding the swap chain as necessary.