Vulkan tutorial progress: textured quad
We now render a textured quad, using anisotropic filtering. The texture is currently hardcoded. I've modified the dimensions of the quad to respect the aspect ratio of the texture file.

As part of the process of following this tutorial, several helper functions have been created and are now used in other parts of the 3D setup process. There are a number of optimizations I can try to make once I understand Vulkan enough to lay everything out to meet my specific needs (ability to render a variable number of unrelated 3D scenes, etc.), so once I'm able to get Vulkan doing the basic things I need, I should search for all "TODO" notes.