Vulkan refactor progress: multiple objects in a scene
Lots of incomplete rewrite attempts in the `vulkan` folder, all of which are currently excluded from the VS project. I'm gonna try and do things gradually, on the DovahKitVulkanSubsystem (hereafter: DKVS).

The DKVS no longer renders a single model hardcoded as a class member; it instead now renders multiple models (hardcoded into the CPP file), each with its own transform and texture. Currently, each texture gets its own VkDeviceMemory, but eventually we'll build a custom heap to fix this.

The fragment shader now takes an array of textures and the vertex shader, an array of object data (currently just transforms). Each shader also takes a "push constant" containing the current texture and object-data indices to use; this push constant is a small struct that can be rapidly updated as we draw each individual object. The use of arrays means that we can use just one descriptor definition for all textures, and update individual entries using `VkWriteDescriptorSet` just as we do when setting the list up initially; for object-data, we can update the storage buffers for any yet-to-be-rendered frames. In essence, we've given our shaders the ability to render multiple objects.

(As an implementation detail, we now use a single texture sampler for all textures passed to the shader, rather than passing texture-and-sampler combos as was done with the single-texture shader.)

Of course, because we aren't adjusting the models' transforms every frame, they no longer spin. We could try updating them every frame (requiring careful writes to the shared buffer; it may be the next thing worth trying), or we could try having the camera orbit the scene center.

As an additional improvement, each rendered object now uses a single VkBuffer for both its vertices and its indices.