Vulkan progress: animations; only syncing object transforms when needed
Removed the original test mesh, test texture, and associated code and fields. Removed the now-unused model transform from the UBO.

Shader modules are now set up once and retained for the life of the singleton, so we don't need to reload compiled shader bytecode every time we rebuild the swap chain.

All rendered objects now have an associated animation state, used to apply a spin animation to our test objects. The UI has buttons that can pause or resume each object's spin.

We now only sync a rendered object's shader parameters if those parameters are "dirty" for each frame. They're marked as "dirty" when changed through accessors; we retain a flags mask with one bit per frame. (This does mean that a swap chain with more than 32 images will work improperly.) Rendered objects have to be marked as "dirty" when the swap chain is rebuilt, so that we do an initial sync for them.