Vulkan: descriptor definitions are easier; variable-length texture array
The fragment shader now takes a variable-length array of texture descriptors. Because this list is variable-length, it must now be the last binding in its containing descriptor set. (The same constraint does not apply to the variable-length list of object-data used by the vertex shader because it is, technically, just one descriptor, a buffer, that we're subdividing.)

As part of the above, we now need the "descriptor indexing" device extension; we need to check whether the features we need from it are supported for a given physical device, and even if they are, we need to explicitly request access to those features when setting up the logical device.

We now have class definitions, `descriptor_binding` and `descriptor_set_layout`, which allow us to describe our desired bindings and sets in one place. We define this data in DovahKitVulkanSubsystem's constructor, now, and then use these structures to automatically determine what parameters to pass when creating the descriptor set layout, the descriptor pool, and the descriptor binding allocations. As such, we no longer have to keep two or three different pieces of code in sync when maintaining our descriptors.

Removed the unused model transform from our uniform buffer object.

Updated our function for updating the object transform shader-side buffer: we can now run vkMapMemory only on the portion of the buffer that we're actually updating. Does that benefit us at all? I have no idea, but it seems like it would, somehow, and it's not much more work.

We now have a separate function for updating animation state.

Added constexpr configuration to the top of the file, to control how often the Vulkan view widget attempts to repaint. It's currently configured to repaint as fast as possible; previously it was hardcoded to repaint at 60 FPS.