Vulkan: cleanup; null texture; adding/removing objects/textures on the fly
The renderer is now capable of using a dummy "null texture" to clear entries in the shaders' texture arrays, when the hardware doesn't support intentionally writing VK_NULL_HANDLE into the arrays.

The renderer no longer requires anisotropic filtering; if the feature is unavailable on the current hardware, it simply won't be used.

The act of resetting and refilling a command buffer is now separate from the act of initially creating the command buffer.

It is now possible to add new objects (and textures) to the scene, and to remove them. Moreover, multiple `rendered_object`s can now share the same `loaded_texture`, referring to it by index and updating its refcount.

It is now possible to destroy objects and textures, removing them from the scene while taking care not to delete them out from any pending frames in flight.

The previous two changes required the addition of a "dirty flags" mask to objects and textures, indicating the frames in flight for which the object or texture needs to be updated. When the object or texture has not been marked for delete, the mask indicates that an update is needed (push the texture into the texture-array descriptor binding; update the object-properties buffer for the vertex shader). When the object or texture has been marked for deletion, the mask indicates which frames in flight may still be using the object, and we can delete the object or texture only when all relevant bits are cleared.

We check whether textures need descriptor updates or descriptor removals whenever we draw a frame; both changes will require a command buffer change (updating descriptors will invalidate a command buffer that had them bound). We check whether objects need updates whenever we draw a frame, but can only remove objects after resetting and refilling the frame's command buffer; we mark all command buffers as invalid (using CPU-side bools) when removing an object. Command buffers are checked for invalidity (and reset and refilled if necessary) when we draw a frame, after we handle any texture and object updates (since, as mentioned, those can invalidate the command buffers).