Vulkan: Start of attempted refactor
Right now, our renderer is a big pile of components, each roughly at the same level and each centrally managed by a singleton. That's not ideal. We can't make every component perfectly self-contained, but I would at least like to try and impose a top-down hierarchy where lower components only refer upward to the uppermost two nodes (a per-surface "context" object or a wrapper for the logical device).

Equally, I want to decouple a number of objects from the swap chain. Based on the reading I've done, it should be possible to have a variety of objects exist per frame-in-flight rather than per swap chain image; similarly, the swap chain extent is more a property of the surface than of the swap chain, and so can be moved to the "context" object.

There are some immediate benefits to this: some objects, like the descriptor pool, can now be set up wholly independently of the swap chain, and need not be rebuilt with the swap chain. Additionally, scenes are easier to manage: we can know at compile-time the right size of the bitmask to use for "dirty" flags on meshes and textures.

I'm not doing a refactor in-place; I'm basically building a separate renderer, and one that ideally won't be locked to a specific widget. It should be possible to create a new `context` for each widget you want to render to.