Work continues on the Vulkan refactor
We now have a QObject wrapper for VkInstance.

What I've come up with so far completely sucks, but it sucks in an educational way. The renderer is now organized in a loose hierarchy, but objects reach into ancestors' and descendants' innards quite frequently, when my intention was for the only upward references to be to the root. Moreover, access to commonly-needed resources on the root (i.e. the VkDevice) is cumbersome.

I initially believed that the render surface could be meaningfully separated from the logical device. This is almost true... true enough to seem completely true. Accordingly, the refactor has a `surface` object which bridges a QWidget to the Vulkan instance, a `logical_device` object, and a `surface_renderer` object which carries out the actual rendering. In reality, we can combine `logical_device` into `surface_renderer`, and we may even be able to combine `surface` into `surface_renderer` (or maybe the other way 'round -- call the result `surface`). We can probably merge `scene` in as well.