Vulkan refactor: program now builds; ready for testing
Modified the main window's header so that it no longer directly includes the render window; hopefully now, we won't risk having to recompile 60% of the entire program whenever we change just the stuff for the render window.

Merged `surface` into `surface_renderer`, and made `DKVulkanInstance` a singleton.

Instances of `physical_device` are now heap-allocated.

The project now builds again, though the new renderer is untested.