Vulkan refactor: fixes; currently stuck on synchronicity issues
Added a test menu item for the new renderer.

Fixed how vertex bindings and attributes are specified in material definitions, and fixed defaults for rasterization, multisampling, and triangle rendering properties.

Fixed a mistake in `buffer` that caused us to lose the buffer's size when move-assigning or move-constructing. Fixed backwards size checks in `buffer::copy_from`.

Fixed crashes in the surface renderer stemming from us attempting to construct a logical device immediately upon being given a widget (or being constructed with one), rather than waiting until a physical device is set. Fixed mistakes stemming from materials being set up before the render pass they depended on. Added a missing (default) color blend to the current test material.

Fixed bad logic for getting swap chain images when rebuilding the swap chain (we forgot to acquire new handles when the image count didn't change). Before tearing down the swap chain, we now use vkDeviceWaitIdle so that all pending render operations can complete.

Fixed several bugs and assertion failures in the new renderer.

The new renderer no longer crashes the program; however, it also fails to render properly (after resizing the window so that the view isn't zero-size, anyway). We get Vulkan validation errors stating that the swap chain image hasn't been transitioned to the right layout at the time we attempt to present it. The image layout transition we need should happen automatically when the render subpass is complete, which suggests a race condition; stepping through a frame-draw in our debugger confirms this (we get the error if we mash through super quickly, but not if we go slowly).

This means, then, that we're not properly making sure that a swap chain image is rendered to *and presented* before attempting to render to it again.